rss-scrapers/pyproject.toml
[tool.poetry]
name = "rss-scrapers"
version = "0.1.0"
description = "RSS Feed Scrapers for AI Trends"
authors = ["Your Name"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
feedparser = "6.0.10"
requests = "2.28.2"
pymongo = "^4.10.1"
beautifulsoup4 = "4.12.0"
python-dateutil = "2.8.2"
pytz = "2023.3"
jupyter = "^1.1.1"
ipykernel = "^6.29.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
